Fix preemtively exiting openssl dtls server for tests

+++ openssl/test/test.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
+if test "$TRAVIS_OS_NAME" == "osx"; then
+	return
+fi
+
 trap "trap - SIGTERM && kill -- -$$" SIGINT SIGTERM EXIT
 
 openssl s_server -accept 15418 -www -cert openssl/test/cert.pem -key openssl/test/key.pem >/dev/null 2>&1 &
@@ -8,9 +12,7 @@ for port in `seq 15411 15430`; do
 	  -key openssl/test/key.pem 2>&1 >/dev/null &
 done
 
-rm -f /tmp/rust_openssl_test_pipe
-mkfifo /tmp/rust_openssl_test_pipe
 # the server for the test ssl::tests::test_write_dtlsv1 must wait to receive
 # data from the client
-openssl s_server -accept 15410 -dtls1 -cert openssl/test/cert.pem \
-  -key openssl/test/key.pem 2>&1 >/dev/null </tmp/rust_openssl_test_pipe
+yes | openssl s_server -accept 15410 -dtls1 -cert openssl/test/cert.pem \
+  -key openssl/test/key.pem 2>&1 >/dev/null
